Panic when building a Struct with Union field, if using Repr
When a Struct `Alpha` has a field `Beta`, and `Beta` is a union, trying to construct `Alpha.Repr` leads to a panic. This is because `Alpha.Repr` tries to assign to `Beta` by treating it as a `UnionRepr`. `AssignNode` calls `AssembleKey` on this `UnionRepr`, which currently panic's since it is not implemented.

Unsure if this is:
1) just a missing implementation (if so, what should the implementation do)
2) a bug and a missing implementation. *Should* UnionRepr.Assemble be getting called, or should it be Union.Assemble? Is this code the Beta node has already been constructed and all it needs to do is be copied over.